# docs/swagger.yaml
definitions:
  model.AnswerJson:
    properties:
      examID:
        type: integer
      img:
        items:
          type: string
        type: array
    type: object
  model.ClassJson:
    properties:
      name:
        type: string
      teacherID:
        type: integer
    type: object
  model.CutMg:
    properties:
      cut:
        items:
          type: string
        type: array
      examID:
        type: string
    type: object
  model.Exam:
    properties:
      createTime:
        type: integer
      desc:
        type: string
      id:
        type: integer
      name:
        type: string
      state:
        type: integer
      subjects:
        type: string
    type: object
  model.PaperJson:
    properties:
      examID:
        type: integer
      imgs:
        items:
          type: string
        type: array
      maxMark:
        items:
          type: integer
        type: array
      ownerID:
        type: integer
    type: object
...
info:
  contact: {}
paths:
  /api/c/all:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 获取所有班级ID
      tags:
      - Class
  /api/c/info:
    post:
      parameters:
      - description: 班级ID列表
        in: body
        name: classIDs
        required: true
        schema:
          items:
            type: integer
          type: array
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 获取班级信息
      tags:
      - Class
  /api/c/mg/add:
    post:
      parameters:
      - description: 班级验证码
        in: formData
        name: classToken
        required: true
        type: string
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 加入班级
      tags:
      - Class
  /api/c/mg/clear:
    post:
      parameters:
      - description: 班级ID
        in: formData
        name: classID
        required: true
        type: integer
      - description: 班级ID
        in: formData
        name: studentID
        required: true
        type: integer
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 将学生请出班级
      tags:
      - Class
  /api/c/mg/quit:
    post:
      parameters:
      - description: 班级ID
        in: formData
        name: classID
        required: true
        type: integer
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 离开班级
      tags:
      - Class
  /api/c/mg/score:
    post:
      parameters:
      - description: 班级ID
        in: formData
        name: classID
        required: true
        type: string
      - description: 考试ID
        in: formData
        name: examID
        required: true
        type: integer
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 查询班级某次考试的成绩
      tags:
      - Class
  /api/c/mg/sh:
    post:
      parameters:
      - description: 班级ID
        in: formData
        name: classID
        required: true
        type: string
      - description: 查询的数目
        in: formData
        name: "n"
        required: true
        type: integer
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 查询班级最近n次的考试成绩
      tags:
      - Class
  /api/c/new:
    post:
      parameters:
      - description: 班级信息
        in: body
        name: classMg
        required: true
        schema:
          $ref: '#/definitions/model.ClassJson'
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 创建班级
      tags:
      - Class
  /api/c/rc:
    post:
      parameters:
      - description: 班级信息
        in: body
        name: classMg
        required: true
        schema:
          $ref: '#/definitions/model.ClassJson'
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 修改班级信息
      tags:
      - Class
  /api/exam/all:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 获取所有考试ID
      tags:
      - Exam
  /api/exam/delete:
    post:
      parameters:
      - description: 考试ID
        in: formData
        name: examID
        required: true
        type: integer
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 删除考试
      tags:
      - Exam
  /api/exam/info:
    post:
      parameters:
      - description: 考试ID列表
        in: body
        name: examIDs
        required: true
        schema:
          items:
            type: integer
          type: array
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 获取考试信息
      tags:
      - Exam
  /api/exam/mg/join:
    post:
      parameters:
      - description: 班级ID
        in: formData
        name: classID
        required: true
        type: integer
      - description: 考试ID
        in: formData
        name: examID
        required: true
        type: integer
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 班级加入考试
      tags:
      - Exam
  /api/exam/mg/num:
    post:
      parameters:
      - description: 数据类型 student teacher task
        in: formData
        name: type
        required: true
        type: string
      - description: 考试ID
        in: formData
        name: examID
        required: true
        type: integer
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 获取考试相关的数据
      tags:
      - Exam
  /api/exam/mg/plan:
    post:
      parameters:
      - description: 改卷任务分段点数组
        in: body
        name: cut
        required: true
        schema:
          $ref: '#/definitions/model.CutMg'
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 分发考试任务
      tags:
      - Exam
  /api/exam/mg/quit:
    post:
      parameters:
      - description: 班级ID
        in: formData
        name: classID
        required: true
        type: integer
      - description: 考试ID
        in: formData
        name: examID
        required: true
        type: integer
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 班级退出考试
      tags:
      - Exam
  /api/exam/mg/score:
    post:
      parameters:
      - description: 考试ID
        in: formData
        name: examID
        required: true
        type: integer
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 获取某一考试所有人的考试成绩
      tags:
      - Exam
  /api/exam/new:
    post:
      parameters:
      - description: 考试信息
        in: body
        name: exam
        required: true
        schema:
          $ref: '#/definitions/model.Exam'
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 创建考试
      tags:
      - Exam
  /api/exam/re:
    post:
      parameters:
      - description: 考试信息
        in: body
        name: exam
        required: true
        schema:
          $ref: '#/definitions/model.Exam'
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 修改考试信息
      tags:
      - Exam
  /api/paper/all:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 获取所有试卷的ID
      tags:
      - Paper
  /api/paper/delete:
    post:
      parameters:
      - description: 试卷ID
        in: formData
        name: paperID
        required: true
        type: integer
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 删除试卷
      tags:
      - Paper
  /api/paper/info:
    post:
      parameters:
      - description: 试卷ID列表
        in: body
        name: paperIDs
        required: true
        schema:
          items:
            type: integer
          type: array
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 获取试卷信息
      tags:
      - Paper
  /api/paper/na:
    post:
      parameters:
      - description: 试卷信息列表
        in: body
        name: answers
        required: true
        schema:
          $ref: '#/definitions/model.AnswerJson'
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 添加试卷答案
      tags:
      - Paper
  /api/paper/new:
    post:
      parameters:
      - description: 试卷信息列表
        in: body
        name: papers
        required: true
        schema:
          items:
            $ref: '#/definitions/model.PaperJson'
          type: array
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 添加试卷
      tags:
      - Paper
  /api/paper/sum:
    post:
      parameters:
      - description: 试卷ID
        in: body
        name: paperIDs
        required: true
        schema:
          items:
            type: integer
          type: array
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 更新试卷的总分
      tags:
      - Paper
...
swagger: "2.0"